Fixed Reservation Lookup Functionality
Created the reservation_lookup.php which would replace the Reservation_Search.php and the Search_Function.php that allows users to look up their reservations via email or confirmation id.

This required a change to the Moffat_Bay_Database  (added a view so that it would be easier to poll the required info).